refactor(refactor): :recycle: Refactor for code climate
