+++ socket.js
@@ -153,33 +153,23 @@ io.on('connection', (socket) => {
 // Live Updates with FIBS (myBallClub) Private API
 function liveUpdate() {
 	fs.readFile(__dirname + '/app/json/settings.json', 'utf8', (err, json) => {
-		if(err)
-			return console.error(err);
+		if(err) return console.error(err);
 		const obj = JSON.parse(json);
-		if(obj.fibsStreaming == false)
-			return console.log('FIBS Streaming is disabled');
-		if(obj.fibsStreamingCode == '')
-			return console.log('FIBS Streaming Code is empty');
+		if(obj.fibsStreaming == false) return console.log('FIBS Streaming is disabled');
+		if(obj.fibsStreamingCode == '') return console.log('FIBS Streaming Code is empty');
 		const IDfibs = obj.fibsStreamingCode;
 		let lastPlay = 1;
 		// Open the JSON file containing the last play
 		fs.readFile('lastPlay.json', (err, data) => {
 			if (err){
 				// Create the file if it does not exist
-				fs.writeFile('lastPlay.json', '0', (err) => {
-					if (err) throw err;
-					console.log('lastPlay.json created');
-				});
+				fs.writeFile('lastPlay.json', '0', (err) => { if (err) throw err; console.log('lastPlay.json created')}); 
+				data = '0';
 			}
 			lastPlay = JSON.parse(data);
 		});
 		// Make the request to the FIBS API
-		const req_option = {
-			hostname: 's3-eu-west-1.amazonaws.com',
-			port: 443,
-			path: '/game.wbsc.org/gamedata/mbc/' + IDfibs +'t.json',
-			method: 'GET'
-		};
+		const req_option = { hostname: 's3-eu-west-1.amazonaws.com', port: 443, path: '/game.wbsc.org/gamedata/mbc/' + IDfibs +'t.json', method: 'GET' };
 		const req = https.request(req_option, (res) => {
 			// If the request is successful update the last play (if it is different from the previous one) and request the new data
 			if (res.statusCode == 200) {
@@ -193,49 +183,32 @@ function liveUpdate() {
 						requestData(IDfibs, data_obj);
 					}
 				});
-			}else{
-				console.log('FIBS update error: ' + res.statusCode);
-			}
+			}else console.log('FIBS update error: ' + res.statusCode);
 		});
 		req.end();
-		req.on('error', (e) => {
-			console.log('FIBS update error: ' + e);
-		});
+		req.on('error', (e) => { console.log('FIBS update error: ' + e) });
 		setTimeout(liveUpdate, 1000);
 	});
 }
 
-function requestData(IDfibs, lastPlay) {
-	// Make the request to the FIBS API
-	const req_option = {
-		hostname: 's3-eu-west-1.amazonaws.com',
-		port: 443,
-		path: '/game.wbsc.org/gamedata/mbc/' + IDfibs +'.json',
-		method: 'GET'
-	};
+function requestData(IDfibs) {
+	const req_option = { hostname: 's3-eu-west-1.amazonaws.com', port: 443, path: '/game.wbsc.org/gamedata/mbc/' + IDfibs +'.json', method: 'GET' };
 	const req = https.request(req_option, (res) => {
 		if(res.statusCode == 200){
 			console.log('FIBS data updated');
 			let data = '';
-			res.on('data', (data2) => {
-				data += data2;
-			});
+			res.on('data', (data2) => { data += data2; });
 			res.on('end', () => {
-				// Update the last play
 				const data_obj = JSON.parse(data);
 				const AwayRuns = data_obj?.awaytotals?.R;
 				const HomeRuns = data_obj?.hometotals?.R;
-				// Remove the index 0 on the arrays
-				const awayRunsParz = data_obj?.awayruns
-				const homeRunsParz = data_obj?.homeruns
+				const { awayruns, homeruns } = data_obj;
 				const bases = { 1: data_obj.runner[1] ? true : false, 2: data_obj.runner[2] ? true : false, 3: data_obj.runner[3] ? true : false, };
 				const inning = parseInt(data_obj.inning);
-				// The arrow is 1 on TOP and 2 on BOT
 				const arrow = data_obj.home == 0 ? 1 : 2;
-				// Destructuring the play and creating the new format for app/json/data.json
 				var int = {};
 				for(let i=1;i<=inning;i++)
-					int[i] = { A: awayRunsParz[i]!=undefined ? awayRunsParz[i]:0, H: homeRunsParz[i]!=undefined ? homeRunsParz[i]:0 }
+					int[i] = { A: awayruns[i]!=undefined ? awayruns[i]:0, H: homeruns[i]!=undefined ? homeruns[i]:0 }
 				fs.readFile(__dirname + '/app/json/data.json', (err,data)=>{
 					if(err) throw err;
 					const oldData = JSON.parse(data);
@@ -248,18 +221,11 @@ function requestData(IDfibs, lastPlay) {
 			});
 		}else{
 			console.log('FIBS data update error: ' + res.statusCode);
-			let chunk = '';
-			res.on('data', (data) => {
-				chunk += data;
-			});
-			res.on('end', () => {
-				console.log(chunk);
-			});
+			let chunk = ''; res.on('data', (data) => { chunk += data; });
+			res.on('end', () => { console.log(chunk); });
 		}
 	});
-	req.on('error', (e) => {
-		console.log('FIBS data update error: ' + e);
-	});
+	req.on('error', (e) => { console.log('FIBS data update error: ' + e); });
 	req.end();
 }
